style(tests): clear clippy::pedantic debt in test code, enforce via --all-targets
Test code (inline #[cfg(test)] modules and tests/) carried 88 pedantic
warnings the CI command did not catch because it omitted --all-targets.
Fixed all of them and added --all-targets to the documented verification
pipeline so test code stays at the same standard as the library:

- Add reasons to 42 bare #[ignore] attributes in real_devices.rs
- Inline format args, digit separators, raw-string hash cleanup (clippy --fix)
- Replace float == with approx::assert_abs_diff_eq in http_integration.rs
- Struct-update syntax instead of field reassign after Default
- u16::try_from path instead of `as u16` cast in time.rs test
- Scoped, justified allow(similar_names) for the x/x_clone test idiom

+++ src/types/time.rs
@@ -497,9 +497,9 @@ mod tests {
 
     #[test]
     fn wakeup_duration_valid_values() {
-        for v in [1, 60, 300, 1800, 3000] {
-            let duration = WakeupDuration::new(Duration::from_secs(v)).unwrap();
-            assert_eq!(duration.seconds(), v as u16);
+        for v in [1u16, 60, 300, 1800, 3000] {
+            let duration = WakeupDuration::new(Duration::from_secs(u64::from(v))).unwrap();
+            assert_eq!(duration.seconds(), v);
         }
     }
